Retyping from float to double

+++ src/com/FuzzyLogic/Main.java
@@ -5,14 +5,14 @@ import java.util.Scanner;
 
 public class Main {
 
-    public static float getCrispInput(String placeholder) {
+    public static double getCrispInput(String placeholder) {
         Scanner input_stream = new Scanner(System.in);
-        float user_input;
+        double user_input;
 
         while(true) {
             try {
                 System.out.print(placeholder + ": ");
-                user_input = input_stream.nextFloat();
+                user_input = input_stream.nextDouble();
                 return user_input;
             }
             catch(Exception InputMismatchException) {
@@ -28,8 +28,8 @@ public class Main {
         System.out.println("===             Using Fuzzy Logic             ===");
         System.out.println("=================================================\n");
 
-        float project_fund = getCrispInput("Project Fund");
-        float experience_level = getCrispInput("Experience Level");
+        double project_fund = getCrispInput("Project Fund");
+        double experience_level = getCrispInput("Experience Level");
 
         System.out.println(project_fund + " " + experience_level);
